# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type JobPriority = typeof JobPriority[keyof typeof JobPriority];


export const JobPriority = {
  URGENT: 'URGENT',
  HIGH: 'HIGH',
  NORMAL: 'NORMAL',
  LOW: 'LOW',
} as const;

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  NEW: 'NEW',
  CONTACTED: 'CONTACTED',
  CONFIRMED: 'CONFIRMED',
  GOING: 'GOING',
  ARRIVED: 'ARRIVED',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  ON_HOLD: 'ON_HOLD',
  POSTPONED: 'POSTPONED',
  CANCELLED: 'CANCELLED',
  UNREACHABLE: 'UNREACHABLE',
} as const;

export type ContactMethod = typeof ContactMethod[keyof typeof ContactMethod];


export const ContactMethod = {
  CALL: 'CALL',
  SMS: 'SMS',
  WHATSAPP: 'WHATSAPP',
  IN_PERSON: 'IN_PERSON',
} as const;

export interface Contact {
  id: number;
  jobId: number;
  date: string;
  time: string;
  method: ContactMethod;
  result: string;
  /** @nullable */
  notes?: string | null;
}

export interface Job {
  id: number;
  jobId: string;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  customerAddress?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  area?: string | null;
  /** @nullable */
  installationDate?: string | null;
  /** @nullable */
  outletName?: string | null;
  /** @nullable */
  agentName?: string | null;
  /** @nullable */
  dealCode?: string | null;
  /** @nullable */
  shopPickName?: string | null;
  /** @nullable */
  shopPickLocation?: string | null;
  /** @nullable */
  deviceNumber?: string | null;
  /** @nullable */
  deviceImei?: string | null;
  /** @nullable */
  deviceType?: string | null;
  /** @nullable */
  jobSource?: string | null;
  /** @nullable */
  dateReceived?: string | null;
  /** @nullable */
  assignedDate?: string | null;
  /** @nullable */
  scheduledDate?: string | null;
  priority: JobPriority;
  status: JobStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  internalNotes?: string | null;
  /** @nullable */
  contactStatus?: string | null;
  /** @nullable */
  lastContactDate?: string | null;
  /** @nullable */
  nextFollowUpDate?: string | null;
  /** @nullable */
  postponementDate?: string | null;
  /** @nullable */
  postponementReason?: string | null;
  /** @nullable */
  completionDate?: string | null;
  /** @nullable */
  completionTime?: string | null;
  transportCost?: number;
  otherExpenses?: number;
  /** @nullable */
  expenseDescription?: string | null;
  totalExpenses?: number;
  /** @nullable */
  customerConfirmation?: string | null;
  /** @nullable */
  installationNotes?: string | null;
  contacts?: Contact[];
}

/**
 * Job fields accepted when creating a job; server-owned id and contacts are ignored.
 */
export type JobInput = Job;

export type JobUpdatePriority = typeof JobUpdatePriority[keyof typeof JobUpdatePriority];


export const JobUpdatePriority = {
  URGENT: 'URGENT',
  HIGH: 'HIGH',
  NORMAL: 'NORMAL',
  LOW: 'LOW',
} as const;

export type JobUpdateStatus = typeof JobUpdateStatus[keyof typeof JobUpdateStatus];


export const JobUpdateStatus = {
  NEW: 'NEW',
  CONTACTED: 'CONTACTED',
  CONFIRMED: 'CONFIRMED',
  GOING: 'GOING',
  ARRIVED: 'ARRIVED',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  ON_HOLD: 'ON_HOLD',
  POSTPONED: 'POSTPONED',
  CANCELLED: 'CANCELLED',
  UNREACHABLE: 'UNREACHABLE',
} as const;

export interface JobUpdate {
  customerName?: string;
  customerPhone?: string;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  customerAddress?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  area?: string | null;
  /** @nullable */
  installationDate?: string | null;
  /** @nullable */
  scheduledDate?: string | null;
  priority?: JobUpdatePriority;
  status?: JobUpdateStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  internalNotes?: string | null;
  /** @nullable */
  contactStatus?: string | null;
  /** @nullable */
  nextFollowUpDate?: string | null;
  /** @nullable */
  postponementDate?: string | null;
  /** @nullable */
  postponementReason?: string | null;
  /** @nullable */
  completionDate?: string | null;
  /** @nullable */
  completionTime?: string | null;
  transportCost?: number;
  otherExpenses?: number;
  /** @nullable */
  expenseDescription?: string | null;
  totalExpenses?: number;
  /** @nullable */
  customerConfirmation?: string | null;
  /** @nullable */
  installationNotes?: string | null;
  [key: string]: unknown;
 }

export type ContactInputMethod = typeof ContactInputMethod[keyof typeof ContactInputMethod];


export const ContactInputMethod = {
  CALL: 'CALL',
  SMS: 'SMS',
  WHATSAPP: 'WHATSAPP',
  IN_PERSON: 'IN_PERSON',
} as const;

export interface ContactInput {
  method: ContactInputMethod;
  result: string;
  notes?: string;
}

export type DashboardByStatus = {[key: string]: number};

export interface Dashboard {
  total: number;
  today: number;
  completedToday: number;
  remainingToday: number;
  followUps: number;
  awaitingResponse: number;
  byStatus: DashboardByStatus;
}

export type SettingsSmsTemplates = {[key: string]: string};

export type SettingsWhatsappTemplates = {[key: string]: string};

export interface Settings {
  id: number;
  name: string;
  phone: string;
  whatsapp: string;
  currency: string;
  defaultStatus: string;
  defaultPriority: string;
  smsTemplates: SettingsSmsTemplates;
  whatsappTemplates: SettingsWhatsappTemplates;
}

export type SettingsInput = Settings;

export type ReportsByStatus = {[key: string]: number};

export type ReportsByAgent = {[key: string]: number};

export type ReportsByArea = {[key: string]: number};

export interface Reports {
  byStatus: ReportsByStatus;
  byAgent: ReportsByAgent;
  byArea: ReportsByArea;
  transportExpenses: number;
  totalExpenses: number;
}