# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactMethod } from './contactMethod';

export interface Contact {
  id: number;
  jobId: number;
  date: string;
  time: string;
  method: ContactMethod;
  result: string;
  /** @nullable */
  notes?: string | null;
}