# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactInputMethod } from './contactInputMethod';

export interface ContactInput {
  method: ContactInputMethod;
  result: string;
  notes?: string;
}