# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardByStatus } from './dashboardByStatus';

export interface Dashboard {
  total: number;
  today: number;
  completedToday: number;
  remainingToday: number;
  followUps: number;
  awaitingResponse: number;
  byStatus: DashboardByStatus;
}