# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Contact } from './contact';
import type { JobPriority } from './jobPriority';
import type { JobStatus } from './jobStatus';

export interface Job {
  id: number;
  jobId: string;
  customerName: string;
  customerPhone: string;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  customerAddress?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  area?: string | null;
  /** @nullable */
  installationDate?: string | null;
  /** @nullable */
  outletName?: string | null;
  /** @nullable */
  agentName?: string | null;
  /** @nullable */
  dealCode?: string | null;
  /** @nullable */
  shopPickName?: string | null;
  /** @nullable */
  shopPickLocation?: string | null;
  /** @nullable */
  deviceNumber?: string | null;
  /** @nullable */
  deviceImei?: string | null;
  /** @nullable */
  deviceType?: string | null;
  /** @nullable */
  jobSource?: string | null;
  /** @nullable */
  dateReceived?: string | null;
  /** @nullable */
  assignedDate?: string | null;
  /** @nullable */
  scheduledDate?: string | null;
  priority: JobPriority;
  status: JobStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  internalNotes?: string | null;
  /** @nullable */
  contactStatus?: string | null;
  /** @nullable */
  lastContactDate?: string | null;
  /** @nullable */
  nextFollowUpDate?: string | null;
  /** @nullable */
  postponementDate?: string | null;
  /** @nullable */
  postponementReason?: string | null;
  /** @nullable */
  completionDate?: string | null;
  /** @nullable */
  completionTime?: string | null;
  transportCost?: number;
  otherExpenses?: number;
  /** @nullable */
  expenseDescription?: string | null;
  totalExpenses?: number;
  /** @nullable */
  customerConfirmation?: string | null;
  /** @nullable */
  installationNotes?: string | null;
  contacts?: Contact[];
}