# lib/api-zod/src/generated/types/jobInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Job } from './job';

/**
 * Job fields accepted when creating a job; server-owned id and contacts are ignored.
 */
export type JobInput = Job;