# lib/api-zod/src/generated/types/jobUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobUpdatePriority } from './jobUpdatePriority';
import type { JobUpdateStatus } from './jobUpdateStatus';

export interface JobUpdate {
  customerName?: string;
  customerPhone?: string;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  customerAddress?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  area?: string | null;
  /** @nullable */
  installationDate?: string | null;
  /** @nullable */
  scheduledDate?: string | null;
  priority?: JobUpdatePriority;
  status?: JobUpdateStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  internalNotes?: string | null;
  /** @nullable */
  contactStatus?: string | null;
  /** @nullable */
  nextFollowUpDate?: string | null;
  /** @nullable */
  postponementDate?: string | null;
  /** @nullable */
  postponementReason?: string | null;
  /** @nullable */
  completionDate?: string | null;
  /** @nullable */
  completionTime?: string | null;
  transportCost?: number;
  otherExpenses?: number;
  /** @nullable */
  expenseDescription?: string | null;
  totalExpenses?: number;
  /** @nullable */
  customerConfirmation?: string | null;
  /** @nullable */
  installationNotes?: string | null;
  [key: string]: unknown;
 }