# lib/api-zod/src/generated/types/reports.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReportsByAgent } from './reportsByAgent';
import type { ReportsByArea } from './reportsByArea';
import type { ReportsByStatus } from './reportsByStatus';

export interface Reports {
  byStatus: ReportsByStatus;
  byAgent: ReportsByAgent;
  byArea: ReportsByArea;
  transportExpenses: number;
  totalExpenses: number;
}