# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SettingsSmsTemplates } from './settingsSmsTemplates';
import type { SettingsWhatsappTemplates } from './settingsWhatsappTemplates';

export interface Settings {
  id: number;
  name: string;
  phone: string;
  whatsapp: string;
  currency: string;
  defaultStatus: string;
  defaultPriority: string;
  smsTemplates: SettingsSmsTemplates;
  whatsappTemplates: SettingsWhatsappTemplates;
}