# lib/api-zod/src/generated/types/settingsInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Settings } from './settings';

export type SettingsInput = Settings;